requirements.txt
behave==1.2.6
parse==1.12.0
parse-type==0.4.2
PyYAML==5.1
selenium==3.141.0
six==1.12.0
urllib3==1.25.3
